backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
bcrypt==4.1.2
PyPDF2==3.0.1
groq==0.4.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.3.2 